PostSolve b2ContactImpulse use raw pointer; webBinding/Box2DWebBindings.cpp cleanup

+++ src/dynamics/b2_contact.cpp
@@ -246,6 +246,6 @@ void b2Contact::Update(b2ContactListener* listener)
 
 	if (sensor == false && touching && listener)
 	{
-		listener->PreSolve((uint32)this, &oldManifold);
+		listener->PreSolve((uint32)this, uint32(&oldManifold));
 	}
 }
